Fixed quoting problem on table insert
As we add the optional regexp for the content match
in the tag field, it could happen that special regexp
characters causes the DB insert to fail with a syntax
error. The PostgreSQL docu says:

you can always use $$ DOLLAR QUOTED STRING LITERALS to
get around shell quoting escaping.

Which is done in this commit